fix: move handler de erro de validacao para classe separada do controller
